+++ noether_tpp/include/noether_tpp/utils.h
@@ -1,5 +1,9 @@
 #include <noether_tpp/core/tool_path_modifier.h>
+
 #include <pcl/PCLPointField.h>
+#include <pcl/geometry/mesh_traits.h>
+#include <pcl/geometry/triangle_mesh.h>
+#include <pcl/PolygonMesh.h>
 
 namespace noether
 {
@@ -18,5 +22,17 @@ std::vector<pcl::PCLPointField>::const_iterator findField(const std::vector<pcl:
 
 std::vector<pcl::PCLPointField>::const_iterator findFieldOrThrow(const std::vector<pcl::PCLPointField>& fields,
                                                                  const std::string& name);
+
+Eigen::Vector3f getPoint(const pcl::PCLPointCloud2& cloud, const std::uint32_t pt_idx);
+
+Eigen::Vector3f getNormal(const pcl::PCLPointCloud2& cloud, const std::uint32_t pt_idx);
+
+using MeshTraits = pcl::geometry::DefaultMeshTraits<std::uint32_t>;
+using TriangleMesh = pcl::geometry::TriangleMesh<MeshTraits>;
+
+/**
+ * @brief Creates a triangle mesh representation of the input polygon mesh
+ */
+TriangleMesh createTriangleMesh(const pcl::PolygonMesh& input);
 
 }  // namespace noether
